+++ ethereum/spec/src/property-test/java/tech/pegasys/teku/spec/datastructures/blobs/versions/deneb/BlindedBlobSidecarsPropertyTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.datastructures.blocks.versions.deneb;
+package tech.pegasys.teku.spec.datastructures.blobs.versions.deneb;
 
 import static tech.pegasys.teku.spec.propertytest.util.PropertyTestHelper.assertDeserializeMutatedThrowsExpected;
 import static tech.pegasys.teku.spec.propertytest.util.PropertyTestHelper.assertRoundTrip;
@@ -19,24 +19,23 @@ import static tech.pegasys.teku.spec.propertytest.util.PropertyTestHelper.assert
 import com.fasterxml.jackson.core.JsonProcessingException;
 import net.jqwik.api.ForAll;
 import net.jqwik.api.Property;
-import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlindedBlobSidecar;
-import tech.pegasys.teku.spec.propertytest.suppliers.blocks.versions.deneb.BlindedBlobSidecarSupplier;
+import tech.pegasys.teku.spec.propertytest.suppliers.blocks.versions.deneb.BlindedBlobSidecarsSupplier;
 
-public class BlindedBlobSidecarPropertyTest {
+public class BlindedBlobSidecarsPropertyTest {
 
   @Property
   void roundTrip(
-      @ForAll(supplier = BlindedBlobSidecarSupplier.class)
-          final BlindedBlobSidecar blindedBlobSidecar)
+      @ForAll(supplier = BlindedBlobSidecarsSupplier.class)
+          final BlindedBlobSidecars blindedBlobSidecars)
       throws JsonProcessingException {
-    assertRoundTrip(blindedBlobSidecar);
+    assertRoundTrip(blindedBlobSidecars);
   }
 
   @Property
   void deserializeMutated(
-      @ForAll(supplier = BlindedBlobSidecarSupplier.class)
-          final BlindedBlobSidecar blindedBlobSidecar,
+      @ForAll(supplier = BlindedBlobSidecarsSupplier.class)
+          final BlindedBlobSidecars blindedBlobSidecars,
       @ForAll final int seed) {
-    assertDeserializeMutatedThrowsExpected(blindedBlobSidecar, seed);
+    assertDeserializeMutatedThrowsExpected(blindedBlobSidecars, seed);
   }
 }
