+++ ethereum/spec/src/property-test/java/tech/pegasys/teku/spec/datastructures/blobs/versions/deneb/BlobsSidecarPropertyTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright ConsenSys Software Inc., 2023
+ * Copyright ConsenSys Software Inc., 2022
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.datastructures.blocks.versions.deneb;
+package tech.pegasys.teku.spec.datastructures.blobs.versions.deneb;
 
 import static tech.pegasys.teku.spec.propertytest.util.PropertyTestHelper.assertDeserializeMutatedThrowsExpected;
 import static tech.pegasys.teku.spec.propertytest.util.PropertyTestHelper.assertRoundTrip;
@@ -19,24 +19,20 @@ import static tech.pegasys.teku.spec.propertytest.util.PropertyTestHelper.assert
 import com.fasterxml.jackson.core.JsonProcessingException;
 import net.jqwik.api.ForAll;
 import net.jqwik.api.Property;
-import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.SignedBlindedBlobSidecar;
-import tech.pegasys.teku.spec.propertytest.suppliers.blocks.versions.deneb.SignedBlindedBlobSidecarSupplier;
+import tech.pegasys.teku.spec.propertytest.suppliers.execution.versions.deneb.BlobsSidecarSupplier;
 
-public class SignedBlindedBlobSidecarPropertyTest {
+public class BlobsSidecarPropertyTest {
 
   @Property
-  void roundTrip(
-      @ForAll(supplier = SignedBlindedBlobSidecarSupplier.class)
-          final SignedBlindedBlobSidecar signedBlindedBlobSidecar)
+  void roundTrip(@ForAll(supplier = BlobsSidecarSupplier.class) final BlobsSidecar blobsSidecar)
       throws JsonProcessingException {
-    assertRoundTrip(signedBlindedBlobSidecar);
+    assertRoundTrip(blobsSidecar);
   }
 
   @Property
   void deserializeMutated(
-      @ForAll(supplier = SignedBlindedBlobSidecarSupplier.class)
-          final SignedBlindedBlobSidecar signedBlindedBlobSidecar,
+      @ForAll(supplier = BlobsSidecarSupplier.class) final BlobsSidecar blobsSidecar,
       @ForAll final int seed) {
-    assertDeserializeMutatedThrowsExpected(signedBlindedBlobSidecar, seed);
+    assertDeserializeMutatedThrowsExpected(blobsSidecar, seed);
   }
 }
