+++ ethereum/spec/src/property-test/java/tech/pegasys/teku/spec/datastructures/blobs/versions/deneb/SignedBlobSidecarsPropertyTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.datastructures.blocks.versions.deneb;
+package tech.pegasys.teku.spec.datastructures.blobs.versions.deneb;
 
 import static tech.pegasys.teku.spec.propertytest.util.PropertyTestHelper.assertDeserializeMutatedThrowsExpected;
 import static tech.pegasys.teku.spec.propertytest.util.PropertyTestHelper.assertRoundTrip;
@@ -19,24 +19,23 @@ import static tech.pegasys.teku.spec.propertytest.util.PropertyTestHelper.assert
 import com.fasterxml.jackson.core.JsonProcessingException;
 import net.jqwik.api.ForAll;
 import net.jqwik.api.Property;
-import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.SignedBlindedBlobSidecars;
-import tech.pegasys.teku.spec.propertytest.suppliers.blocks.versions.deneb.SignedBlindedBlobSidecarsSupplier;
+import tech.pegasys.teku.spec.propertytest.suppliers.blocks.versions.deneb.SignedBlobSidecarsSupplier;
 
-public class SignedBlindedBlobSidecarsPropertyTest {
+public class SignedBlobSidecarsPropertyTest {
 
   @Property
   void roundTrip(
-      @ForAll(supplier = SignedBlindedBlobSidecarsSupplier.class)
-          final SignedBlindedBlobSidecars signedBlindedBlobSidecars)
+      @ForAll(supplier = SignedBlobSidecarsSupplier.class)
+          final SignedBlobSidecars signedBlobSidecars)
       throws JsonProcessingException {
-    assertRoundTrip(signedBlindedBlobSidecars);
+    assertRoundTrip(signedBlobSidecars);
   }
 
   @Property
   void deserializeMutated(
-      @ForAll(supplier = SignedBlindedBlobSidecarsSupplier.class)
-          final SignedBlindedBlobSidecars signedBlindedBlobSidecars,
+      @ForAll(supplier = SignedBlobSidecarsSupplier.class)
+          final SignedBlobSidecars signedBlobSidecars,
       @ForAll final int seed) {
-    assertDeserializeMutatedThrowsExpected(signedBlindedBlobSidecars, seed);
+    assertDeserializeMutatedThrowsExpected(signedBlobSidecars, seed);
   }
 }
